backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
bcrypt
pyjwt
python-dotenv
sqlite3